requirements.txt
tkinter
qrcode
Pillow